% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{angles}
\alias{angles}
\alias{rad}
\alias{deg}
\title{Convert between angles and radians}
\usage{
deg(radian)

rad(degree)
}
\arguments{
\item{radian, degree}{angles with different units}
}
\description{
Convert between angles and radians
}
