% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_fits}
\alias{bind_fits}
\title{ModelFit Bind list of \code{\link{ModelFit}} objects into single fit}
\usage{
bind_fits(ModelFits)
}
\arguments{
\item{ModelFits}{list of ModelFit objects}
}
\value{
\code{\link[=deg]{deg()}} converts radian to degree, \code{\link[=rad]{rad()}} converts degress to radian
}
\description{
ModelFit Bind list of \code{\link{ModelFit}} objects into single fit
}
\details{
thin wrapper around \code{\link[rstan:sflist2stanfit]{rstan::sflist2stanfit()}}
}
