% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{cross_threshold}
\alias{cross_threshold}
\title{Get Thresholds}
\usage{
cross_threshold(d, group, x, y, quantiles = c(0, 0.9), participant = NULL)
}
\arguments{
\item{d}{wide,dataframe containing the}

\item{group}{name of column (bare only)}

\item{x, y}{names of column (bare or quoted) in d with which to do the regression}

\item{quantiles}{threshold to apply.}

\item{participant}{optional secondary grouping column. Can be used to nest thresholding
(e.g., calculate the quantiles within each participant separately)}
}
\value{
data frame object with one column for the \code{group}, potentially another index
column for \code{participant}, and a new column called \code{Threshold.} The \code{Threshold}
column is crossed with the original dataset, but for each crossing, members
of \code{group} are excluded if they do not cross the threshold.
}
\description{
Get Thresholds
}
\examples{
sub02 |>
  tidyr::pivot_wider(names_from = contrast, values_from = y) |>
  cross_threshold(voxel, low, high)

# can also calculate within groups
sub02 |>
  tidyr::pivot_wider(names_from = contrast, values_from = y) |>
  cross_threshold(c(voxel, run), low, high, participant = sub)
}
\seealso{
\code{\link[=get_slope]{get_slope()}}, \code{\link[=get_slope_by_group]{get_slope_by_group()}}
}
