% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deming-class.R
\name{Deming}
\alias{Deming}
\title{Model object, ready for sampling}
\description{
\code{Deming} facilitates a non-parametric check of neuromodulation.
This class is a wrapper around a \code{\link[cmdstanr:CmdStanModel]{cmdstanr::CmdStanModel}} class.
}
\examples{

## ------------------------------------------------
## Method `Deming$new`
## ------------------------------------------------

m <- sub02 \%>\%
     tidyr::pivot_wider(names_from = contrast, values_from = y) \%>\%
     dplyr::mutate(orientation = factor(orientation)) \%>\%
     Deming$new(low, high, tuning_var = orientation, voxel_var = voxel)
m

m$cmdstanmodel
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cmdstanr_version}}{Version of cmdstanr used to build models}

\item{\code{cmdstan_version}}{Version of cmdstan used to build models}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{standata}}{used to fit model}

\item{\code{prior}}{used to fit model}

\item{\code{cmdstanmodel}}{Underlying \code{\link[cmdstanr:CmdStanModel]{cmdstanr::CmdStanModel}}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Deming$new()}}
\item \href{#method-make_standata}{\code{Deming$make_standata()}}
\item \href{#method-sample}{\code{Deming$sample()}}
\item \href{#method-clone}{\code{Deming$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize new instance of class Deming
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Deming$new(d, x, y, tuning_var, voxel_var = "voxel", prior = DemingPrior$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{dataframe from which to make standata.}

\item{\code{x, y}}{Names of columns in d which contain the x and y values}

\item{\code{tuning_var}}{Name of column across which there was testing}

\item{\code{voxel_var}}{Name of column indexing voxels}

\item{\code{prior}}{A \code{\link{DemingPrior}}}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{m <- sub02 \%>\%
     tidyr::pivot_wider(names_from = contrast, values_from = y) \%>\%
     dplyr::mutate(orientation = factor(orientation)) \%>\%
     Deming$new(low, high, tuning_var = orientation, voxel_var = voxel)
m

m$cmdstanmodel
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_standata"></a>}}
\if{latex}{\out{\hypertarget{method-make_standata}{}}}
\subsection{Method \code{make_standata()}}{
Prepare data for running model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Deming$make_standata(d, x, y, tuning_var, voxel_var = "voxel")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{dataframe from which to make standata.}

\item{\code{x, y}}{Names of columns in d which contain the x and y values.}

\item{\code{tuning_var}}{Name of column across which there was testing. Column must be a factor.}

\item{\code{voxel_var}}{Name of column indexing voxels. Column must be a factor.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
named list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
Draw samples from the posterior of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Deming$sample(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{arguments passed to \link[cmdstanr:model-method-sample]{cmdstanr::sample()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{\link[cmdstanr:CmdStanMCMC]{cmdstanr::CmdStanMCMC}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Deming$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
