% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{.make_vtf0}
\alias{.make_vtf0}
\title{create voxel tuning function posterior}
\usage{
.make_vtf0(post, cores = 1)
}
\arguments{
\item{post}{output of model}

\item{cores}{integer number of cores over which to parallelize}
}
\value{
array n_draws x n_chains x n_unique obs. for use in loo.VtuneFit
}
\description{
create voxel tuning function posterior
}
