% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total-least-squares.R
\name{get_slope}
\alias{get_slope}
\title{Return slope in total least squares problem}
\usage{
get_slope(x, y)
}
\arguments{
\item{x}{vector with independent variable}

\item{y}{vector with dependent variable}
}
\value{
the slope of the line that minimizes the total least squared error
}
\description{
Return slope in total least squares problem
}
\details{
Whereas ordinary least squares finds the line that minimizes the sum of squared
vertical distance, total least squares finds the line that minimizes the sum
of vertical and horizontal differences.
}
\examples{
n <- 1000
z <- rnorm(n)
x <- rnorm(n, z)
y <- rnorm(n, 2*x + 1)
get_slope(x,y)

}
\seealso{
For a more general function (including one that returns the intercept), see \code{\link[pracma:odregress]{pracma::odregress()}}
}
