% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanmodel.R
\name{get_stanmodel}
\alias{get_stanmodel}
\title{get_stanmodel}
\usage{
get_stanmodel(model_name)
}
\arguments{
\item{model_name}{chracter that is either "deming" or "vtf"}
}
\value{
compiled cmdstanmodel
}
\description{
get_stanmodel
}
\examples{
stanmodels("deming")
stanmodels("vtf")
}
