% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtf.R
\name{make_standata}
\alias{make_standata}
\title{Prepare data for running model}
\usage{
make_standata(d, model, priors = set_priors())
}
\arguments{
\item{d}{dataframe from which to make standata. The output of this can be passed to \code{\link[=vtf]{vtf()}}.}

\item{model}{Form of neuromodulation. One of "multiplicative" or "additive".}

\item{priors}{named list. see output of \code{\link[=set_priors]{set_priors()}}}
}
\description{
Prepare data for running model
}
