% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\name{Model}
\alias{Model}
\title{Model object, ready for sampling}
\description{
Model object, ready for sampling

Model object, ready for sampling
}
\seealso{
\code{\link[=make_standata]{make_standata()}}
}
\section{Super class}{
\code{\link[cmdstanr:CmdStanModel]{cmdstanr::CmdStanModel}} -> \code{Model}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cmdstanr_version}}{Version of cmdstanr used to build models}

\item{\code{cmdstan_version}}{Version of cmdstan used to build models}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sample}{\code{Model$sample()}}
\item \href{#method-clone}{\code{Model$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="cmdstanr" data-topic="CmdStanModel" data-id="check_syntax">}\href{../../cmdstanr/html/CmdStanModel.html#method-check_syntax}{\code{cmdstanr::CmdStanModel$check_syntax()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cmdstanr" data-topic="CmdStanModel" data-id="code">}\href{../../cmdstanr/html/CmdStanModel.html#method-code}{\code{cmdstanr::CmdStanModel$code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cmdstanr" data-topic="CmdStanModel" data-id="compile">}\href{../../cmdstanr/html/CmdStanModel.html#method-compile}{\code{cmdstanr::CmdStanModel$compile()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cmdstanr" data-topic="CmdStanModel" data-id="cpp_options">}\href{../../cmdstanr/html/CmdStanModel.html#method-cpp_options}{\code{cmdstanr::CmdStanModel$cpp_options()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cmdstanr" data-topic="CmdStanModel" data-id="exe_file">}\href{../../cmdstanr/html/CmdStanModel.html#method-exe_file}{\code{cmdstanr::CmdStanModel$exe_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cmdstanr" data-topic="CmdStanModel" data-id="generate_quantities">}\href{../../cmdstanr/html/CmdStanModel.html#method-generate_quantities}{\code{cmdstanr::CmdStanModel$generate_quantities()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cmdstanr" data-topic="CmdStanModel" data-id="hpp_file">}\href{../../cmdstanr/html/CmdStanModel.html#method-hpp_file}{\code{cmdstanr::CmdStanModel$hpp_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cmdstanr" data-topic="CmdStanModel" data-id="initialize">}\href{../../cmdstanr/html/CmdStanModel.html#method-initialize}{\code{cmdstanr::CmdStanModel$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cmdstanr" data-topic="CmdStanModel" data-id="optimize">}\href{../../cmdstanr/html/CmdStanModel.html#method-optimize}{\code{cmdstanr::CmdStanModel$optimize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cmdstanr" data-topic="CmdStanModel" data-id="print">}\href{../../cmdstanr/html/CmdStanModel.html#method-print}{\code{cmdstanr::CmdStanModel$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cmdstanr" data-topic="CmdStanModel" data-id="save_hpp_file">}\href{../../cmdstanr/html/CmdStanModel.html#method-save_hpp_file}{\code{cmdstanr::CmdStanModel$save_hpp_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cmdstanr" data-topic="CmdStanModel" data-id="stan_file">}\href{../../cmdstanr/html/CmdStanModel.html#method-stan_file}{\code{cmdstanr::CmdStanModel$stan_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cmdstanr" data-topic="CmdStanModel" data-id="variational">}\href{../../cmdstanr/html/CmdStanModel.html#method-variational}{\code{cmdstanr::CmdStanModel$variational()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
Draw samples from the posterior of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$sample(data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data to pass directly to stan. Can be produced with \code{\link[=make_standata]{make_standata()}}.}

\item{\code{...}}{further arguments passed to \code{\link[cmdstanr:model-method-sample]{cmdstanr::model-method-sample()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
