% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfit-class.R
\name{ModelFit}
\alias{ModelFit}
\title{ModelFit object, output from sampling}
\description{
ModelFit object, output from sampling

ModelFit object, output from sampling
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rawdata}}{The original data that was used to prepare the model}

\item{\code{prior}}{The \code{\link{Prior}} that was used to fit the model}

\item{\code{stanfit}}{\code{\link[rstan:stanfit-class]{rstan::stanfit}} output}

\item{\code{model}}{\code{\link{Model}} object used to generate the fit}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelFit$new()}}
\item \href{#method-standata}{\code{ModelFit$standata()}}
\item \href{#method-loo}{\code{ModelFit$loo()}}
\item \href{#method-make_vtf0}{\code{ModelFit$make_vtf0()}}
\item \href{#method-.lfun}{\code{ModelFit$.lfun()}}
\item \href{#method-.llfun}{\code{ModelFit$.llfun()}}
\item \href{#method-clone}{\code{ModelFit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new ModelFit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFit$new(rawdata, prior, model, stanfit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rawdata}}{dataframe from which to make standata}

\item{\code{prior}}{\code{\link{Prior}} that was used to fit the model}

\item{\code{model}}{\code{\link{Model}} used to draw samples}

\item{\code{stanfit}}{\code{\link[rstan:stanfit-class]{rstan::stanfit}} output by model}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ModelFit} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standata"></a>}}
\if{latex}{\out{\hypertarget{method-standata}{}}}
\subsection{Method \code{standata()}}{
Reconstruct the list of data that was passed to Stan
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFit$standata()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list
Estimate the LOO score
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loo"></a>}}
\if{latex}{\out{\hypertarget{method-loo}{}}}
\subsection{Method \code{loo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFit$loo(cores = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cores}}{int}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{loo::psis_loo} object
create voxel tuning function posterior
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_vtf0"></a>}}
\if{latex}{\out{\hypertarget{method-make_vtf0}{}}}
\subsection{Method \code{make_vtf0()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFit$make_vtf0(cores = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cores}}{integer number of cores over which to parallelize}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
array n_draws x n_chains x n_unique obs. for use in \code{\link[=ModelFit$loo]{ModelFit$loo()}}
Likelihood function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.lfun"></a>}}
\if{latex}{\out{\hypertarget{method-.lfun}{}}}
\subsection{Method \code{.lfun()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFit$.lfun(data_i, draws)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_i}}{single row of data}

\item{\code{draws}}{posterior}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The likelihood of a single observation
Log-Likelihood function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.llfun"></a>}}
\if{latex}{\out{\hypertarget{method-.llfun}{}}}
\subsection{Method \code{.llfun()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFit$.llfun(data_i, draws)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_i}}{single row of data}

\item{\code{draws}}{posterior}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The log-likelihood of a single observation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
