% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelmcmc-class.R
\name{ModelMCMC}
\alias{ModelMCMC}
\title{ModelFit object, output from sampling}
\description{
ModelFit object, output from sampling

ModelFit object, output from sampling
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{standata}}{The original data that was used to prepare the model}

\item{\code{cmdstanmcmc}}{\code{\link[cmdstanr:CmdStanMCMC]{CmdStanMCMC}} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelMCMC$new()}}
\item \href{#method-loo}{\code{ModelMCMC$loo()}}
\item \href{#method-make_vtf0}{\code{ModelMCMC$make_vtf0()}}
\item \href{#method-.lfun}{\code{ModelMCMC$.lfun()}}
\item \href{#method-.llfun}{\code{ModelMCMC$.llfun()}}
\item \href{#method-clone}{\code{ModelMCMC$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new ModelFit
You probably don't want to call this
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelMCMC$new(standata, cmdstanmcmc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{standata}}{Data that were used to generate the samples}

\item{\code{cmdstanmcmc}}{\code{\link[cmdstanr:CmdStanMCMC]{cmdstanr::CmdStanMCMC}} output by model}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ModelFit} object.
Estimate the LOO score
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loo"></a>}}
\if{latex}{\out{\hypertarget{method-loo}{}}}
\subsection{Method \code{loo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelMCMC$loo(cores = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cores}}{int}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{loo::psis_loo} object
create voxel tuning function posterior
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_vtf0"></a>}}
\if{latex}{\out{\hypertarget{method-make_vtf0}{}}}
\subsection{Method \code{make_vtf0()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelMCMC$make_vtf0(cores = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cores}}{integer number of cores over which to parallelize}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[posterior:draws_array]{posterior::draws_array}},
Likelihood function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.lfun"></a>}}
\if{latex}{\out{\hypertarget{method-.lfun}{}}}
\subsection{Method \code{.lfun()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelMCMC$.lfun(data_i, draws)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_i}}{single row of data}

\item{\code{draws}}{posterior}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The likelihood of a single observation
Log-Likelihood function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.llfun"></a>}}
\if{latex}{\out{\hypertarget{method-.llfun}{}}}
\subsection{Method \code{.llfun()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelMCMC$.llfun(data_i, draws)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_i}}{single row of data}

\item{\code{draws}}{posterior}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The log-likelihood of a single observation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelMCMC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
