% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmmr-package.R
\docType{package}
\name{nmmr-package}
\alias{nmmr-package}
\alias{_PACKAGE}
\alias{nmmr}
\title{nmmr.}
\description{
Implement NeuroModulation Modelling in R
}
\seealso{
Useful links:
\itemize{
  \item \url{https://psadil.github.io/nmmr/}
  \item \url{https://github.com/psadil/nmmr}
  \item Report bugs at \url{https://github.com/psadil/nmmr/issues}
}

}
\author{
\strong{Maintainer}: Patrick Sadil \email{psadil@gmail.com} (\href{https://orcid.org/0000-0003-4141-1343}{ORCID})

}
\keyword{internal}
