% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{Prior}
\alias{Prior}
\title{List of priors}
\description{
List of priors

List of priors
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sigma_loc}}{\eqn{\sigma_{loc}} ~ \code{gamma(sigma_loc[1], sigma_loc[2])}}

\item{\code{sigma_scale}}{\eqn{\sigma_{scale}} ~ \code{gamma(sigma_scale[1], sigma_scale[2])}}

\item{\code{gamma_loc}}{\eqn{\gamma_{loc}} ~ \code{normal(0, gamma_loc[2])}}

\item{\code{gamma_scale}}{\eqn{\gamma_{scale}} ~ \code{gamma(gamma_scale[1], gamma_scale[2])}}

\item{\code{kappa_loc}}{\eqn{\kappa_{loc}} ~ \code{gamma(kappa_loc[1], kappa_loc[2])}}

\item{\code{kappa_scale}}{\eqn{\kappa_{scale}} ~ \code{gamma(kappa_scale[1], kappa_scale[2])}}

\item{\code{alpha_loc}}{\eqn{\alpha_{loc}} ~ \code{normal(alpha_loc[1], alpha_loc[2])}}

\item{\code{alpha_scale}}{\eqn{\alpha_{scale}} ~ \code{gamma(alpha_scale[1], alpha_scale[2])}}

\item{\code{ntfp_loc}}{\eqn{ntfp_{loc}} ~ \code{normal(ntfp_loc[1], ntfp_loc[2])}}

\item{\code{ntfp_scale}}{\eqn{ntfp_{scale}} ~ \code{gamma(ntfp_scale[1], ntfp_scale[2])}}

\item{\code{prior_only}}{bool indicating whether to include the likelihood (1 => prior predictive check)}

\item{\code{prior_only}}{bool indicating whether to include the likelihood (1 => prior predictive check)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Prior$new()}}
\item \href{#method-as_list}{\code{Prior$as_list()}}
\item \href{#method-clone}{\code{Prior$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Model object.
All values must be greater than 0.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Prior$new(
  sigma_loc = c(2, 1/2),
  sigma_scale = c(2, 1/2),
  gamma_loc = 5,
  gamma_scale = c(2, 1/20),
  kappa_loc = c(3, 1),
  kappa_scale = c(3, 1),
  alpha_loc = c(0, 5),
  alpha_scale = c(2, 1/2),
  ntfp_loc = 0.5,
  ntfp_scale = c(2, 3),
  prior_only = 0,
  sample_yrep = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sigma_loc}}{vector of length 2}

\item{\code{sigma_scale}}{vector of length 2}

\item{\code{gamma_loc}}{real}

\item{\code{gamma_scale}}{vector of length 2}

\item{\code{kappa_loc}}{vector of length 2}

\item{\code{kappa_scale}}{vector of length 2}

\item{\code{alpha_loc}}{vector of length 2}

\item{\code{alpha_scale}}{vector of length 2}

\item{\code{ntfp_loc}}{real}

\item{\code{ntfp_scale}}{vector of length 2}

\item{\code{prior_only}}{0 or 1 indicating whether to include the likelihood (1 => prior predictive check)}

\item{\code{sample_yrep}}{0 or 1 indicating whether to simulate from the posterior}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Prior} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
Convert object to a list.
This is useful when passing data to Stan
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Prior$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Prior$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
