% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{ranks}
\alias{ranks}
\alias{build_ranks}
\alias{visualize_ranks}
\title{Prepare and visualize rank-ordered beta values}
\usage{
build_ranks(d, quantiles = c(0, 0.5, 0.75, 0.9))

visualize_ranks(ranks)
}
\arguments{
\item{d}{data.frame, like \link{sub02}}

\item{quantiles}{Values at which to filter voxels}

\item{ranks}{As output from \code{\link[=build_ranks]{build_ranks()}}}
}
\value{
\code{\link[=build_ranks]{build_ranks()}} constructs a data.frame which can be passed on to \code{\link[=visualize_ranks]{visualize_ranks()}}

\code{\link[=visualise_ranks]{visualise_ranks()}} uses that output to make a basic \code{\link[ggplot2:ggplot2-package]{ggplot2}} figure
}
\description{
Prepare and visualize rank-ordered beta values
}
