% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtf.R
\name{set_priors}
\alias{set_priors}
\title{Set priors for model}
\usage{
set_priors(
  prior_sigma_loc = c(2, 1/2),
  prior_sigma_scale = c(2, 1/2),
  prior_gamma_loc = 5,
  prior_gamma_scale = c(2, 1/20),
  prior_kappa_loc = c(3, 1),
  prior_kappa_scale = c(3, 1),
  prior_alpha_loc = c(0, 5),
  prior_alpha_scale = c(2, 1/2),
  prior_ntfp_loc = 0.5,
  prior_ntfp_scale = c(2, 3)
)
}
\arguments{
\item{prior_sigma_loc}{vector of length 2}

\item{prior_sigma_scale}{vector of length 2}

\item{prior_gamma_loc}{real}

\item{prior_gamma_scale}{vector of length 2}

\item{prior_kappa_loc}{vector of length 2}

\item{prior_kappa_scale}{vector of length 2}

\item{prior_alpha_loc}{vector of length 2}

\item{prior_alpha_scale}{vector of length 2}

\item{prior_ntfp_loc}{real}

\item{prior_ntfp_scale}{vector of length 2}
}
\value{
named list
}
\description{
Set priors for model
}
\details{
all values must be greater than 0

\eqn{\sigma_loc ~ gamma(prior_sigma_loc[1], prior_sigma_loc[2])}

\eqn{\sigma_scale ~ gamma(prior_sigma_scale[1], prior_sigma_scale[2])}

\eqn{\gamma_loc ~ normal(0, prior_gamma_loc[2])}

\eqn{\gamma_scale ~ gamma(prior_gamma_scale[1], prior_gamma_scale[2])}

\eqn{\kappa_loc ~ gamma(prior_kappa_loc[1], prior_kappa_loc[2])}

\eqn{\kappa_scale ~ gamma(prior_kappa_scale[1], prior_kappa_scale[2])}

\eqn{\alpha_loc ~ normal(prior_alpha_loc[1], prior_alpha_loc[2])}

\eqn{\alpha_scale ~ gamma(prior_alpha_scale[1], prior_alpha_scale[2])}

\eqn{ntfp_loc ~ normal(prior_ntfp_loc[1], prior_ntfp_loc[2])}

\eqn{ntfp_scale ~ gamma(prior_ntfp_scale[1], prior_ntfp_scale[2])}
}
\examples{

set_priors()
set_priors(prior_sigma_loc = c(2, 1))
}
