% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VtuneFit-class.R
\name{standata}
\alias{standata}
\title{extract data that was passed to model}
\usage{
standata(object)
}
\arguments{
\item{object}{object to grab rawdata from.}
}
\description{
extract data that was passed to model
}
\details{
see output of \code{\link[=make_standata]{make_standata()}}
}
