% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonal.R
\name{sum_squares}
\alias{sum_squares}
\title{Sum of Squared Error}
\usage{
sum_squares(x, y = x)
}
\arguments{
\item{x}{numeric vector}

\item{y}{optional, numeric vector, same length as \code{x}}
}
\value{
either the sum of squared errors for \code{x} (if \code{y} is left out),
or the cross-product of two vectors \code{x} with \code{y}
}
\description{
Sum of Squared Error
}
\details{
The sum of squared errors for a length \eqn{n} vector \eqn{x} will equal

\deqn{S_{xx} = \sum_i^n(x - \overline{x})^2,}

where \eqn{\overline{x}} is the mean of \eqn{x}.

The cross-product is

\deqn{S_{xy} = \sum_i^n (x - \overline{x})(y - \overline{y})}
}
\examples{
# the sum of squared errors for x
x <- rnorm(10)
sum_squares(x)

# a cross-product
y <- rnorm(10)
sum_squares(x,y)
}
