% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtf.R
\name{vtf}
\alias{vtf}
\title{Run One of the Stan Models}
\usage{
vtf(model, priors, standata = NULL, d = NULL, ...)
}
\arguments{
\item{model}{Form of neuromodulation. One of "multiplicative" or "additive"}

\item{priors}{Vector of doubles.}

\item{...}{see \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}
}
\value{

}
\description{
Run One of the Stan Models
}
\seealso{
\code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}
}
