% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VtuneFit-class.R, R/loo.R
\docType{class}
\name{VtuneFit-class}
\alias{VtuneFit-class}
\alias{VtuneFit}
\alias{rawdata,VtuneFit-method}
\alias{standata,VtuneFit-method}
\alias{bind_fits,VtuneFit-method}
\alias{loo,VtuneFit-method}
\title{VtuneFit inherits from \link[rstan:stanfit-class]{rstan::stanfit} in the \pkg{rstan} package}
\usage{
\S4method{rawdata}{VtuneFit}(object)

\S4method{standata}{VtuneFit}(object)

\S4method{bind_fits}{VtuneFit}(sfitlist)

\S4method{loo}{VtuneFit}(x, ..., cores = 1)
}
\arguments{
\item{object}{object to grab data from.}

\item{sfitlist}{list of VtuneFit objects}

\item{x}{VtuneFit object, containing parameters "mu" and "sigma"}

\item{cores}{int}
}
\description{
VtuneFit inherits from \link[rstan:stanfit-class]{rstan::stanfit} in the \pkg{rstan} package
}
\details{
checkout \link[rstan:stanfit-class]{rstan::stanfit}
and maybe \href{http://rstudio.com}{Rstudio}
}
\section{Methods (by generic)}{
\itemize{
\item \code{rawdata}: extract data used to construct \code{\link[=rawdata]{rawdata()}}

\item \code{standata}: extract data passed to stan. can be reused as input to \code{\link[=vtf]{vtf()}}

\item \code{bind_fits}: Bind list of VtuneFit objects into single fit

\item \code{loo}: Leave-one-out cross-validation
}}

\section{Slots}{

\describe{
\item{\code{rawdata}}{tbl_df. see example dataset, \link{sub02}}

\item{\code{standata}}{list.}
}}

