% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtuner-package.R
\docType{package}
\name{vtuner-package}
\alias{vtuner-package}
\alias{_PACKAGE}
\alias{vtuner}
\title{The 'vtuner' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
