% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{WISEsummary}
\alias{WISEsummary}
\title{WISEsummmary}
\usage{
WISEsummary(
  data,
  dependentvars,
  betweenvars = NULL,
  withinvars = NULL,
  idvar = NULL,
  CI_width = 0.95,
  na.rm = FALSE
)
}
\arguments{
\item{data}{A data frame}

\item{dependentvars}{Character vector giving the dependent variable}

\item{betweenvars}{Character vector giving the between subject variables}

\item{withinvars}{Character vector giving the within subject variables}

\item{idvar}{Character vector giving the name of the column holding subject
identifiers}

\item{CI_width}{Numeric vector giving the confidence level for computing the
confidence interval boundaries. Must be between 0 and 1, non-inclusive.}

\item{na.rm}{a logical value indicating whether NA values should be removed from the Dependent Variables.}
}
\value{
A data frame with summary statistics
}
\description{
Within-subject Error Summary
}
\examples{

WISEsummary(sub02, y, withinvars = c(contrast, orientation), idvar = voxel)
}
\references{
\href{http://www.cookbook-r.com/Graphs/Plotting_means_and_error_bars_(ggplot2)/}{WISESummary},
\href{http://tqmp.org/Content/vol04-2/p061/p061.html}{Morey (2008)}
}
