% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VtuneFit-class.R
\name{bind_fits}
\alias{bind_fits}
\title{Bind list of VtuneFit objects into single fit}
\usage{
bind_fits(sfitlist)
}
\arguments{
\item{sfitlist}{list of VtuneFit objects}
}
\description{
Bind list of VtuneFit objects into single fit
}
\details{
thin wrapper around \code{\link[rstan:sflist2stanfit]{rstan::sflist2stanfit()}}
}
