% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonal.R
\name{get_slope}
\alias{get_slope}
\title{Return slope in orthogonal regression problem}
\usage{
get_slope(x, y)
}
\arguments{
\item{x}{vector with independent variable}

\item{y}{vector with dependent variable}
}
\value{
the slope of the line that minimizes the total least squared error
}
\description{
Return slope in orthogonal regression problem
}
\details{
Whereas ordinary least squares finds the line that minimizes the sum of squared
vertical distance, orthogonal regressions finds the line that minimizes the sum
of vertical and horizontal differences.
}
\examples{
n <- 1000
z <- rnorm(n)
x <- rnorm(n, z)
y <- rnorm(n, 2*z + 1)
get_slope(x,y)
}
\seealso{
\code{\link[=sum_squares]{sum_squares()}}, \code{\link[=get_slope_by_group]{get_slope_by_group()}}
For a more general function (including one that returns the intercept), see \href{https://rdrr.io/rforge/pracma/man/odregress.html}{\code{pracma::odregress()}}
}
