% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonal.R
\name{get_slope_by_group}
\alias{get_slope_by_group}
\title{Calculate slopes by groups}
\usage{
get_slope_by_group(d, group, x, y)
}
\arguments{
\item{d}{wide,dataframe containing the}

\item{group}{name of column (bare only)}

\item{x, y}{names of column (bare or quoted) in d with which to do the regression}
}
\value{
a dataframe with two columns, one containing the group id and
one containing the estimated slope for that group
}
\description{
Calculate slopes by groups
}
\examples{
sub02 \%>\%
  tidyr::pivot_wider(names_from = contrast, values_from = y) \%>\%
  get_slope_by_group(voxel, low, high)

# create group with intersection of variables
sub02 \%>\%
  tidyr::pivot_wider(names_from = contrast, values_from = y) \%>\%
  get_slope_by_group(c(voxel, sub), low, high)

}
\seealso{
\code{\link[=get_slope]{get_slope()}}, \code{\link[=cross_threshold]{cross_threshold()}}
}
