% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standata.R
\name{make_standata}
\alias{make_standata}
\title{Prepare data for running model}
\usage{
make_standata(d, form, prior = Prior$new())
}
\arguments{
\item{d}{dataframe from which to make standata. The output of this can be passed to \code{\link[=vtf]{vtf()}}.}

\item{form}{Form of the neuromodulation (additive or multiplicative)}

\item{prior}{A \code{\link{Prior}}}
}
\value{
named list
}
\description{
Prepare data for running model
}
