% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtf.R
\name{make_standata}
\alias{make_standata}
\title{Make list of data for fitting}
\usage{
make_standata(d, model, priors)
}
\arguments{
\item{d}{tbl_df (see betas)}

\item{model}{name}

\item{priors}{vector}
}
\description{
Make list of data for fitting
}
