% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\name{Model}
\alias{Model}
\title{Model object, ready for sampling}
\description{
Model object, ready for sampling

Model object, ready for sampling
}
\seealso{
\code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{form}}{Form of the neuromodulation (additive or multiplicative)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Model$new()}}
\item \href{#method-sample}{\code{Model$sample()}}
\item \href{#method-make_standata}{\code{Model$make_standata()}}
\item \href{#method-clone}{\code{Model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Model object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$new(form)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{form}}{Form of neuromodulation in the model.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Model} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
Draw samples from the posterior of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$sample(d = NULL, standata = NULL, prior = Prior$new(), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{dataframe from which to make standata}

\item{\code{standata}}{data to pass directly to stan. Can be produced with \code{\link[=make_standata]{make_standata()}}. If present,
this data will be used rather than the data stored in \code{d}.}

\item{\code{prior}}{An object of class \code{\link{Prior}}}

\item{\code{...}}{further arguments passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{\link{ModelFit}}.
Prepare data for running model
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_standata"></a>}}
\if{latex}{\out{\hypertarget{method-make_standata}{}}}
\subsection{Method \code{make_standata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$make_standata(d, prior = Prior$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{dataframe from which to make standata. The output of this can be passed to \code{\link[=vtf]{vtf()}}.}

\item{\code{prior}}{A \code{\link{Prior}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
named list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
