% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\name{Model}
\alias{Model}
\title{Model object, ready for sampling}
\description{
Model object, ready for sampling

Model object, ready for sampling
}
\seealso{
\link[cmdstanr:model-method-sample]{cmdstanr::sample()}
Prepare data for running model
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cmdstanr_version}}{Version of cmdstanr used to build models}

\item{\code{cmdstan_version}}{Version of cmdstan used to build models}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{standata}}{used to fit model}

\item{\code{form}}{used to fit model}

\item{\code{prior}}{used to fit model}

\item{\code{cmdstanmodel}}{Underlying \code{\link[cmdstanr:CmdStanModel]{cmdstanr::CmdStanModel}}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Model$new()}}
\item \href{#method-sample}{\code{Model$sample()}}
\item \href{#method-make_standata}{\code{Model$make_standata()}}
\item \href{#method-clone}{\code{Model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$new(d, form, prior = Prior$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{dataframe from which to make standata.}

\item{\code{form}}{Form of the neuromodulation (additive or multiplicative)}

\item{\code{prior}}{A \code{\link{Prior}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
Draw samples from the posterior of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$sample(data = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{optional new data. If present, this will be used instead
of the internally stored data (created when the object was initialized)}

\item{\code{...}}{arguments passed to \link[cmdstanr:model-method-sample]{cmdstanr::sample()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_standata"></a>}}
\if{latex}{\out{\hypertarget{method-make_standata}{}}}
\subsection{Method \code{make_standata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$make_standata(d)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{dataframe from which to make standata.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
named list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
