% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmmr-package.R
\docType{package}
\name{nmmr-package}
\alias{nmmr-package}
\alias{_PACKAGE}
\alias{nmmr}
\title{The 'nmmr' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
\author{
\strong{Maintainer}: Patrick Sadil \email{psadil@gmail.com} (\href{https://orcid.org/0000-0003-4141-1343}{ORCID})

}
\keyword{internal}
