% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub02.R
\docType{data}
\name{sub02}
\alias{sub02}
\title{Sample beta values}
\format{
Objects of class tbl_df (inherits from tbl, data.frame) a minimal tibble for the \code{nmmr} analysis (53856 x 7)

\describe{
\item{sub}{factor ID of subject}
\item{run}{factor ID of session}
\item{voxel}{factor ID of voxel}
\item{contrast}{Factor with different values of contrast. Note that the factor \emph{must} have the lower levels of contrast correspond
to a lower integer representation (e.g., low == 1 and high == 2).}
\item{orientation}{Orientation presented on trial, in radians (between \eqn{\pm\pi}).
This is the \emph{tuning variable}. Most functions require the numeric values, but some functions require only the factor levels (e.g., \link[=Deming]{Deming$new()} ).}
\item{y}{beta values in given condition (corresponds to beta in betas2)}
\item{ses}{factor ID of session}
}
}
\usage{
sub02
}
\value{
a \link[tibble:tbl_df-class]{tibble}
}
\description{
Example dataset for testing analyses, and to see input structure
}
\details{
Where possible, the columns of this dataframe follow the \href{https://bids.neuroimaging.io/}{BIDS}
format. That is, \code{sub} specifies participant, and \code{run} is the run number
within session \code{ses}.

The main functions of this package require data formatted analogously. At a
minimum, you will need a column indexing voxels (e.g., \code{voxel}, in \code{sub02}),
a column for the dependent variable (e.g., \code{y}), a column indexing the stimulus
values for which voxels might be tuned (e.g., \code{orientation}), and a column
indexing the experimental factor across which the tuning may be modulated
(e.g., \code{contrast}). A column indexing participant (e.g., \code{sub}) is required
for running the full hierarchical versions of the models.  The other columns
(e.g., \code{run} and \code{ses}) are not always necessary, but they help keep the data
tidy and may also help with troubleshooting (e.g., voxel activity often differs
across sessions).
}
\examples{
sub02
}
\keyword{datasets}
