% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtf.R
\name{vtf}
\alias{vtf}
\title{Run One of the Stan Models}
\usage{
vtf(model, d = NULL, standata = NULL, priors = set_priors(), ...)
}
\arguments{
\item{model}{Form of neuromodulation. One of "multiplicative" or "additive".}

\item{d}{dataframe from which to make standata. The output of this can be passed to \code{\link[=vtf]{vtf()}}.}

\item{standata}{data to pass directly to stan. Can be produced with \code{\link[=make_standata]{make_standata()}}. If present,
this data will be used rather than the data stored in \code{d}.}

\item{priors}{named list. see output of \code{\link[=set_priors]{set_priors()}}.}

\item{...}{further arguments passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}.}
}
\value{

}
\description{
Run One of the Stan Models
}
\seealso{
\code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}
}
